+++ scriptabit/__init__.py
@@ -16,7 +16,14 @@ from .habitica_service import HabiticaService, HabiticaTaskTypes
 from .habitica_task import HabiticaTask
 from .habitica_task_service import HabiticaTaskService
 from .iplugin import IPlugin
-from .scriptabit import start_cli
+from .scriptabit import (
+    start_scriptabit,
+    start_banking,
+    start_csv,
+    start_health,
+    start_pets,
+    start_trello,
+)
 from .task import (
     Task,
     Difficulty,